Refactor GPBoost example and model: integrate SyntheticSleepDataGenerator, implement dataset creation function, and update GPBoostTimeSeriesModel to accept label tokenizer.
